refactor: Split linutil into TUI and Core crates
