+++ tabs/applications-setup/alacritty-setup.sh
