+++ tabs/applications-setup/kitty-setup.sh
