+++ tabs/applications-setup/rofi-setup.sh
