+++ tabs/applications-setup/zsh-setup.sh
