+++ tabs/common-script.sh
