+++ tabs/system-setup/1-compile-setup.sh
