+++ tabs/system-setup/2-gaming-setup.sh
