+++ tabs/system-setup/4-remove-snaps.sh
