+++ tabs/system-setup/arch/paru-setup.sh
