+++ tabs/system-setup/arch/server-setup.sh
