+++ tabs/system-setup/arch/yay-setup.sh
