+++ tabs/system-setup/fedora/rpm-fusion-setup.sh
