+++ tabs/utils/monitor-control/auto_detect_displays.sh
