+++ tabs/utils/monitor-control/disable_monitor.sh
