+++ tabs/utils/monitor-control/duplicate_displays.sh
