+++ tabs/utils/monitor-control/enable_monitor.sh
