+++ tabs/utils/monitor-control/extend_displays.sh
