+++ tabs/utils/monitor-control/manage_arrangement.sh
