+++ tabs/utils/monitor-control/reset_scaling.sh
