+++ tabs/utils/monitor-control/scale_monitor.sh
