+++ tabs/utils/monitor-control/set_resolutions.sh
