+++ tabs/utils/monitor-control/utility_functions.sh
