+++ tui/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "linutil_tui"
+edition = "2021"
+version.workspace = true
+license.workspace = true
+
+[dependencies]
+clap = { version = "4.5.16", features = ["derive"] }
+crossterm = "0.27.0"
+ego-tree = { workspace = true }
+oneshot = "0.1.8"
+portable-pty = "0.8.1"
+ratatui = "0.27.0"
+tui-term = "0.1.12"
+unicode-width = "0.1.13"
+linutil_core = { path = "../core" }
+
+[build-dependencies]
+chrono = "0.4.33"
+
+[[bin]]
+name = "linutil"
+path = "src/main.rs"
