+++ tui/src/running_command.rs
@@ -1,5 +1,6 @@
 use crate::float::FloatContent;
 use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
+use linutil_core::Command;
 use oneshot::{channel, Receiver};
 use portable_pty::{
     ChildKiller, CommandBuilder, ExitStatus, MasterPty, NativePtySystem, PtySize, PtySystem,
@@ -13,7 +14,6 @@ use ratatui::{
 };
 use std::{
     io::Write,
-    path::PathBuf,
     sync::{Arc, Mutex},
     thread::JoinHandle,
 };
@@ -21,13 +21,6 @@ use tui_term::{
     vt100::{self, Screen},
     widget::PseudoTerminal,
 };
-
-#[derive(Clone, Hash, Eq, PartialEq)]
-pub enum Command {
-    Raw(String),
-    LocalFile(PathBuf),
-    None, // Directory
-}
 
 pub struct RunningCommand {
     /// A buffer to save all the command output (accumulates, until the command exits)
